+++ src/realm/impl/transact_log.hpp
@@ -334,15 +334,12 @@ public:
     template <class InstructionHandler>
     void parse(util::InputStream&, InstructionHandler&);
 
-    template <class InstructionHandler>
-    void parse(util::NoCopyInputStream&, InstructionHandler&);
-
 private:
     util::Buffer<char> m_input_buffer;
 
     // The input stream is assumed to consist of chunks of memory organised such that
     // every instruction resides in a single chunk only.
-    util::NoCopyInputStream* m_input;
+    util::InputStream* m_input;
     // pointer into transaction log, each instruction is parsed from m_input_begin and onwards.
     // Each instruction are assumed to be contiguous in memory.
     const char* m_input_begin;
@@ -678,20 +675,13 @@ inline TransactLogParser::~TransactLogParser() noexcept {}
 
 
 template <class InstructionHandler>
-void TransactLogParser::parse(util::NoCopyInputStream& in, InstructionHandler& handler)
+void TransactLogParser::parse(util::InputStream& in, InstructionHandler& handler)
 {
     m_input = &in;
     m_input_begin = m_input_end = nullptr;
 
     while (has_next())
         parse_one(handler); // Throws
-}
-
-template <class InstructionHandler>
-void TransactLogParser::parse(util::InputStream& in, InstructionHandler& handler)
-{
-    util::NoCopyInputStreamAdaptor in_2(in, m_input_buffer);
-    parse(in_2, handler); // Throws
 }
 
 inline bool TransactLogParser::has_next() noexcept